+++ dalesbred-junit/src/test/java/fi/evident/dalesbred/junit/TransactionalTestsWithProviderTest.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright (c) 2012 Evident Solutions Oy
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package fi.evident.dalesbred.junit;
+
+import fi.evident.dalesbred.Database;
+import fi.evident.dalesbred.TransactionCallback;
+import fi.evident.dalesbred.TransactionContext;
+import org.jetbrains.annotations.NotNull;
+import org.junit.Rule;
+import org.junit.Test;
+
+import javax.inject.Provider;
+
+import static fi.evident.dalesbred.Propagation.MANDATORY;
+import static org.junit.Assert.assertEquals;
+
+public class TransactionalTestsWithProviderTest {
+
+    private final Database db = TestDatabaseProvider.databaseForProperties("connection.properties");
+
+    @Rule
+    public final TransactionalTests tx = new TransactionalTests(new Provider<Database>() {
+        @Override
+        public Database get() {
+            return db;
+        }
+    });
+
+    @Test
+    public void checkThatThereIsAnExistingTransaction() {
+        // The following code fails if we don't have an active transaction.
+        String result = db.withTransaction(MANDATORY, new TransactionCallback<String>() {
+            @NotNull
+            @Override
+            public String  execute(@NotNull TransactionContext tx) {
+                return "ok";
+            }
+        });
+
+        assertEquals("ok", result);
+    }
+}